+++ ios/Sources/DKMadsSSPSDK/Core/SDKVersion.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public let SDK_VERSION = "0.5.9"
+public let SDK_VERSION = "0.5.10"
 
 public enum SDKError: Error, LocalizedError {
     case notInitialized